main-5

+++ MyBankServer/MyBank/Controllers/CreditRequestsController.cs
@@ -1,5 +1,7 @@
 ﻿using System.Diagnostics.Eventing.Reader;
+using System.Xml;
 using Microsoft.AspNetCore.Cors;
+using Microsoft.EntityFrameworkCore.Query;
 using MyBank.API.DataTransferObjects.CreditRequestDtos;
 
 namespace MyBank.API.Controllers;
@@ -11,18 +13,32 @@ public class CreditRequestsController : ControllerBase
 {
     private readonly ICreditRequestsService _creditRequestsService;
     private readonly IMapper _mapper;
+    private readonly ICookieValidator _cookieValidator;
 
-    public CreditRequestsController(ICreditRequestsService creditRequestsService, IMapper mapper)
+    public CreditRequestsController(ICreditRequestsService creditRequestsService, IMapper mapper, ICookieValidator cookieValidator)
     {
         _creditRequestsService = creditRequestsService;
         _mapper = mapper;
+        _cookieValidator = cookieValidator;
     }
 
     [HttpPost]
     [Authorize(Policy = AuthorizationPolicies.UserPolicy)]
     public async Task<IResult> Add([FromBody] CreditRequestDto dto)
     {
-        var serviceResponse = await _creditRequestsService.Add(_mapper.Map<CreditRequest>(dto));
+        var (status, message, errorCode, role, id) = _cookieValidator.HandleCookie(Request.Headers.Cookie[0]!);
+
+        if (status == false)
+        {
+            return Results.Json(new ErrorDto
+            {
+                ControllerName = "CreditRequestsController",
+                Message = message!
+            },
+            statusCode: 400);
+        }
+
+        var serviceResponse = await _creditRequestsService.Add(dto.Name, dto.CreditPackageId, id!.Value);
 
         if (serviceResponse.Status == false)
         {
@@ -40,9 +56,54 @@ public class CreditRequestsController : ControllerBase
     }
 
     [HttpGet]
-    [Authorize(Policy = AuthorizationPolicies.UserAndModeratorAndAdminPolicy)]
+    [Authorize(Policy = AuthorizationPolicies.ModeratorAndAdminPolicy)]
+    public async Task<IResult> GetById(int creditRequestId, bool includeData)
+    {
+        var (status, message, errorCode, role, id) = _cookieValidator.HandleCookie(Request.Headers.Cookie[0]!);
+
+        if (status == false)
+        {
+            return Results.Json(new ErrorDto
+            {
+                ControllerName = "CreditRequestsController",
+                Message = message!
+            },
+            statusCode: 400);
+        }
+
+        var serviceResponse = await _creditRequestsService.GetById(creditRequestId, includeData);
+
+        if (serviceResponse.Status == false)
+        {
+            return Results.Json(
+                new ErrorDto
+                {
+                    ControllerName = "CreditRequestsController",
+                    Message = serviceResponse.Message,
+                },
+                statusCode: 400
+            );
+        }
+
+        return Results.Json(new { list = MyJsonConverter<CreditRequest>.Convert(serviceResponse.Data) }, statusCode: 200);
+    }
+
+    [HttpGet]
+    [Authorize(Policy = AuthorizationPolicies.ModeratorAndAdminPolicy)]
     public async Task<IResult> GetAllInfoByUserId(int userId)
     {
+        var (status, message, errorCode, role, id) = _cookieValidator.HandleCookie(Request.Headers.Cookie[0]!);
+
+        if (status == false)
+        {
+            return Results.Json(new ErrorDto
+            {
+                ControllerName = "CreditRequestsController",
+                Message = message!
+            },
+            statusCode: 400);
+        }
+
         var serviceResponse = await _creditRequestsService.GetAllByUser(userId);
 
         if (serviceResponse.Status == false)
@@ -57,18 +118,59 @@ public class CreditRequestsController : ControllerBase
             );
         }
 
-        return Results.Json(new { list = serviceResponse.Data }, statusCode: 200);
+        return Results.Json(new { list = MyJsonConverter<List<CreditRequest>>.Convert(serviceResponse.Data) }, statusCode: 200);
+    }
+
+    [HttpGet]
+    [Authorize(Policy = AuthorizationPolicies.ModeratorAndAdminPolicy)]
+    public async Task<IResult> GetAllInfo(bool includeData, bool? isAnswered, bool? isApproved)
+    {
+        var (status, message, errorCode, role, id) = _cookieValidator.HandleCookie(Request.Headers.Cookie[0]!);
+
+        if (status == false)
+        {
+            return Results.Json(new ErrorDto
+            {
+                ControllerName = "CreditRequestsController",
+                Message = message!
+            },
+            statusCode: 400);
+        }
+
+        var serviceResponse = await _creditRequestsService.GetAllInfo(includeData, isAnswered, isApproved);
+
+        if (serviceResponse.Status == false)
+        {
+            return Results.Json(
+                new ErrorDto
+                {
+                    ControllerName = "CreditRequestsController",
+                    Message = serviceResponse.Message,
+                },
+                statusCode: 400
+            );
+        }
+
+        return Results.Json(new { list = MyJsonConverter<List<CreditRequest>>.Convert(serviceResponse.Data) }, statusCode: 200);
     }
 
     [HttpPut]
     [Authorize(Policy = AuthorizationPolicies.ModeratorPolicy)]
-    public async Task<IResult> UpdateStatus(int creditRequestId, int moderatorId, bool isApproved)
+    public async Task<IResult> UpdateState(int creditRequestId, bool isApproved)
     {
-        var serviceResponse = await _creditRequestsService.UpdateIsApproved(
-            creditRequestId,
-            moderatorId,
-            isApproved
-        );
+        var (status, message, errorCode, role, id) = _cookieValidator.HandleCookie(Request.Headers.Cookie[0]!);
+
+        if (status == false)
+        {
+            return Results.Json(new ErrorDto
+            {
+                ControllerName = "CreditRequestsController",
+                Message = message!
+            },
+            statusCode: 400);
+        }
+
+        var serviceResponse = await _creditRequestsService.UpdateIsApproved(creditRequestId, id!.Value, isApproved);
 
         if (serviceResponse.Status == false)
         {
@@ -89,6 +191,18 @@ public class CreditRequestsController : ControllerBase
     [Authorize(Policy = AuthorizationPolicies.UserAndModeratorPolicy)]
     public async Task<IResult> Delete(int creditRequestId)
     {
+        var (status, message, errorCode, role, id) = _cookieValidator.HandleCookie(Request.Headers.Cookie[0]!);
+
+        if (status == false)
+        {
+            return Results.Json(new ErrorDto
+            {
+                ControllerName = "CreditRequestsController",
+                Message = message!
+            },
+            statusCode: 400);
+        }
+
         var serviceResponse = await _creditRequestsService.Delete(creditRequestId);
 
         if (serviceResponse.Status == false)
